Resolved CID 351741 (#1 of 1): Invalid type in argument to printf format specifier
invalid_type: Argument this->m_charters[0].size() to format specifier %u was expected to have type unsigned int but has type unsigned long

+++ src/world/Objects/ObjectMgr.cpp
@@ -2472,7 +2472,7 @@ void ObjectMgr::LoadGuildCharters()
     }
     while (result->NextRow());
     delete result;
-    LogDetail("ObjectMgr : %u charters loaded.", m_charters[0].size());
+    LogDetail("ObjectMgr : %u charters loaded.", static_cast<uint32_t>(m_charters[0].size()));
 }
 
 Charter* ObjectMgr::GetCharter(uint32 CharterId, CharterTypes Type)
